// CryptoVaultRN/src/features/home/bottomTab/explore/explore.type.ts
interface Project {
  endDate: string;
  projectBanner: string;
  projectName: string;
  startDate: string;
  _id: string;
  jettonMasterAddress: string;
  tokenPriceChart?: string;
}

export interface ClaimableNFTProps {}

interface NFTProtcolType {
  logo?: string;
  name?: string;
  slip0044: number;
}

interface ProjectNFTType {
  name?: string;
  amount: number;
  contractAddress: string;
}
type ProjectTokenType = {
  symbol?: string;
  name: string;
  contractAddress: string;
  priceChartScanURL?: string;
};
export interface ClaimableType {
  claimableNFTs?: ClaimableNFTProps[];
  nftCollectionName?: string;
  project?: Project;
  projectNftProtocol?: NFTProtcolType;
  projectNFT: ProjectNFTType;
  projectToken: ProjectTokenType;
  projectNFTs: ProjectNFTType[];
  projectTokenProtocol: {
    _id: string;
    name: string;
    symbol: string;
    blockExplorerUrl: string;
    logo: string;
    transactionScanURL?: string;
  };
}
// ...
export interface OwnedNFTType {
  _id: string;
  amount: number;
  contractAddress: string;
  image: string;
  nftId: string;
  nfts: [];
  rereceiverWalletAddress: string;
  totalAmount: number;
}

interface Project {
  name: string;
  protocolName: string;
}

export interface ClaimHistory {
  _id: string;
  claimGroupId: string;
  nftWalletAddress: string;
  tokenReceiverWalletAddress: string;
  nftId: string;
  amount: number;
  nftContractAddress: string;
  claimableTokenProject: string;
  createdAt: string;
  updatedAt: string;
  __v: number;
}